Reorganise and deduplicate
