+++ Dockerfile
@@ -13,7 +13,7 @@ RUN go mod download
 COPY . .
 
 # Build the application
-RUN CGO_ENABLED=0 GOOS=linux go build -o main ./src/main.go
+RUN CGO_ENABLED=0 GOOS=linux go build -o main ./cmd/app/main.go
 
 # Final stage
 FROM alpine:3.19
