+++ internal/crawler/crawler/types.go
