+++ internal/jobs/queue_helpers.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/Harvey-AU/blue-banded-bee/src/common"
+	"github.com/Harvey-AU/blue-banded-bee/internal/common"
 	"github.com/getsentry/sentry-go"
 	"github.com/rs/zerolog/log"
 )
